Reference

QuadratureOnImplicitRegions.d1_case — Method

d1_case(ψ_list,s_list,domain,x_ref,w_ref)

Returns a quadrature rule on the subset where si*ψi<0 (see the paper).

QuadratureOnImplicitRegions.d1_count_subintervals — Method

d1_count_subintervals(ψ_list,s_list,domain)

Counts how many sub-intervals of domain satisfiy siψi≥0

QuadratureOnImplicitRegions.find_root — Method

find_root(ψ::F,a::T,b::T)

Finds a root of a function ψ in the open interval (a,b)

QuadratureOnImplicitRegions.find_roots — Method

find_roots(ψ::F,a::T,b::T) where {F,T}

Returns multiple roots of ψ in the open interval [a,b]

QuadratureOnImplicitRegions.find_roots — Method

find_roots(ψ_list::Vector{F},a::T,b::T) where {F,T}

Returns multiple roots of multiple functions ψ in the closed interval [a,b], it includes the endpoints by default even if they are not roots.

QuadratureOnImplicitRegions.shifted_gl! — Method

shifted_gl!(x_ref::Vector{T},w_ref::Vector{T},a::T,b::T,x::Vector{T},w::Vector{T})

Shifts the Gauss-Legendre rule (xref,wref) from (-1,1) to the interval [a,b] and stores them in x,w.

QuadratureOnImplicitRegions.shifted_gl — Method

shifted_gl(x_ref::Vector{T},w_ref::Vector{T},a::T,b::T)

Shifts the Gauss-Legendre rule (xref,wref) from (-1,1) to the interval [a,b]

QuadratureOnImplicitRegions.tensor_GL_rule — Method

tensor_GL_rule(U::Matrix{T},x_ref::Vector{T},w_ref::Vector{T})

Returns a matrix x and a vector w (multidimensional quadrature)